#132 checkAndFilterContingencies now adds to a vector timeOfEvent is a float now timeStart_ and durationSimu_ are float now too

+++ sources/Inputs/include/Configuration.h
@@ -121,7 +121,7 @@ class Configuration {
    *
    * @returns the parameter value
    */
-  double getTimeOfEvent() const {
+  float getTimeOfEvent() const {
     return timeOfEvent_;
   }
 
@@ -164,7 +164,7 @@ class Configuration {
   float startTime_ = 0;                                                      ///< Moment (in seconds) at which starts the simulation
   float stopTime_ = 100;                                                      ///< Moment (in seconds) at which ends the simulation
   // Security Analysis only
-  double timeOfEvent_ = 80.0;                                             ///< When will the contengencies be simulated
+  float timeOfEvent_ = 80.0;                                             ///< When will the contengencies be simulated
   int numberOfThreads_ = 4;                                                     ///< The number of threads used to simulate
 };
